[issue25] Handle errors in search.
